Roger setup as a small problem to debug, pushing in hope of getting help debbuging.

+++ apps/roger_lulesh/src/RAG.h
@@ -47,7 +47,7 @@
 
 struct DomainObject_t {
   ocrGuid_t guid;
-  uint8_t  *base;
+  uint64_t *base;
   size_t    offset;
   size_t    limit;
 };
